move performance year graph generation to background job
